sudoers: allow qubes group to set working and root directories
sudo allows the user to specify a working directory (via ‘-D’ or
‘--chdir’) and root directory (via ‘-R’ or ‘--chroot’) for the process
it executes.  This must specifically be enabled in the sudoers file,
which this commit does.

I am marking this as a draft because it *will* break Debian systems, as
their sudo doesn’t support the CWD= and CHROOT= options.  It must not be
merged until we figure out how to strip these options from the Debian
sudoers file.